refactor(map): 抽象出TokenPicker
